pyproject.toml
[project]
name = "uplite"
version = "0.1.0"
description = "A lightweight, configurable web dashboard and connection watcher"
readme = "README.md"
authors = [
    { name = "ppmatrix", email = "ppmatrixcsk@gmail.com" }
]
requires-python = ">=3.12"
keywords = ["dashboard", "monitoring", "flask", "web", "connection-watcher"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Topic :: System :: Monitoring",
    "Topic :: Internet :: WWW/HTTP :: WSGI :: Application",
    "Framework :: Flask",
]
dependencies = [
    "flask>=3.1.1",
    "flask-sqlalchemy>=3.1.1",
    "flask-migrate>=4.1.0",
    "flask-login>=0.6.3",
    "flask-wtf>=1.2.2",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "psutil>=7.0.0",
    "email-validator>=2.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.4.1",
    "pytest-flask>=1.3.0",
    "pytest-cov>=6.2.1",
    "black>=25.1.0",
    "flake8>=7.3.0",
    "mypy>=1.17.1",
]
postgres = [
    "psycopg2-binary>=2.9.0",
]
mysql = [
    "PyMySQL>=1.1.0",
]
redis = [
    "redis>=5.0.0",
]

[project.scripts]
uplite = "uplite:main"

[project.urls]
Homepage = "https://github.com/ppmatrix/uplite"
Repository = "https://github.com/ppmatrix/uplite.git"
Documentation = "https://github.com/ppmatrix/uplite/blob/main/README.md"
"Bug Reports" = "https://github.com/ppmatrix/uplite/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--disable-warnings",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
]

[tool.black]
line-length = 100
target-version = ['py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.mypy]
python_version = "3.12"
check_untyped_defs = true
ignore_missing_imports = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true

[tool.coverage.run]
source = ["src"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]

[tool.flake8]
max-line-length = 100
extend-ignore = ["E203", "W503"]
exclude = [
    ".git",
    ".venv",
    "__pycache__",
    "build",
    "dist",
]
